backend/pyproject.toml
[tool.poetry]
name = "mafia-backend"
version = "0.1.0"
description = "Mafia backend"
authors = ["Core Team Mafia"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.116.1"
uvicorn = {extras=["standard"], version="^0.30.6"}
pydantic = "^2.11.7"
pydantic-settings = "^2.10.1"
SQLAlchemy = "^2.0.43"
asyncpg = "^0.30.0"
alembic = "^1.16.5"
redis = "^6.4.0"
sse-starlette = "^3.0.2"
PyJWT = "^2.10.1"
livekit-api = "^1.0.5"
python-socketio = {version="^5.11.0", extras=["asgi"]}
structlog = "^25.4.0"

[tool.poetry.group.dev.dependencies]

[tool.ruff]
line-length = 100
target-version = "py312"
extend-select = ["I", "UP"]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
strict_optional = true
warn_unused_ignores = true
